Refactor test into folders by subject
Organize the test files
