+++ test/credentials/test_unit_credential_store_factory.py
