+++ test/credentials/test_unit_keystore_adapter.py
