+++ test/credentials/test_unit_keystore_credentials_provider.py
