+++ test/util/test_runners/integrations_tests_runner.py
@@ -35,18 +35,16 @@ import logging
 # NOTE do not delete this import even though it seems to not be used. This used to
 # run the integration tests as a process. Otherwise this dependency will not be available
 # once the test module is packed
-import uuid
-import ssl
 
 # Internals
-from test.test_integration_policy import CliIntegrationPolicy
-from test.test_integration_variable import CliIntegrationTestVariable
+from test.policy.test_integration_policy import CliIntegrationPolicy
+from test.variable.test_integration_variable import CliIntegrationTestVariable
 from test.test_integration_configurations import CliIntegrationTestConfigurations
-from test.test_integration_credentials_keyring import CliIntegrationTestCredentialsKeyring
-from test.test_integration_list import CliIntegrationTestList
-from test.test_integration_credentials_netrc import CliIntegrationTestCredentialsNetrc
-from test.test_integration_oss import CliIntegrationTestOSS
-from test.test_integration_resource import CliIntegrationResourceTest
+from test.credentials.test_integration_credentials_keyring import CliIntegrationTestCredentialsKeyring
+from test.list.test_integration_list import CliIntegrationTestList
+from test.credentials.test_integration_credentials_netrc import CliIntegrationTestCredentialsNetrc
+from test.init.test_integration_oss import CliIntegrationTestOSS
+from test.resource.test_integration_resource import CliIntegrationResourceTest
 from test.util.test_runners.integration_test_case import IntegrationTestCaseBase
 from test.util.test_runners.test_runner_args import TestRunnerArgs
 from test.util.test_runners.params import ClientParams, TestEnvironmentParams
